--- TalkSsogi_server/src/main/java/com/talkssogi/TalkSsogi_server/service/Page7Service.java
@@ -1,7 +1,0 @@
-package com.talkssogi.TalkSsogi_server.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class Page7Service {
-}
